backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
google-api-python-client==2.179.0
google-auth==2.40.3
openai==1.101.0
PyMuPDF==1.26.4
pgvector==0.4.1
python-multipart==0.0.20
